Bug 1676662 - Prepend list bullet when getting range at start of list item. r=morgan
Safari has an inconsistent way to deal with list bullets in text.
If a given range has a list or list item nested inside it - but it is not
at the beginning, the bullet will not be returned as part of the text.But if the range begins at the start
of a list item, the preceding bullet is included. For example, given the following list:

* First Item
* Second Item

If the range is [irst Item], the text for that range should return as "irst Item",
but if the range is [First it], the returned text is "* First it".

This patch emulates that inconsistency by removing the list item considerations
in the text segments iterator, and instead prepending a bullet if needed.

Differential Revision: https://phabricator.services.mozilla.com/D97717

UltraBlame original commit: f62764a6d94d23774ec49cf9b2def38bccbf42dd