Bug 1676909 - Don't evict items from the texture cache if they were used in the previous frame. r=gw
Webrender uses an LRUCache to hold the items in the texture
cache. When texture usage is over a certain threshold we evict the
least recently used items until we are back under the
threshold. However, this runs in to a problem on pages where the cache
only contains items still in use, but is still over the threshold, as
even the least recently used item is still required. In this scenario
we end up evicting items at the start of the frame, only to reupload
them later in the frame, and repeat the cycle again on the next frame.

To avoid this, tweak the eviction algortithm so that it never evicts
items that were in use in the previous frame. (The eviction step
occurs before we know which items are needed for the current frame, so
using the previous frame is the best approximation.)

Differential Revision: https://phabricator.services.mozilla.com/D98043

UltraBlame original commit: 0ee685602a7f4224777b9f1780535c278ad17c0b