Bug 1679409 - [devtools] Write a test for color scheme simulation. r=rcaliman.
The test is quite simple as it checks the state of the buttons as we cycle through
it, and checks that the rule view does show the expected applied rule when
simulating a color scheme.

Even if the color scheme simulation is behind a pref, this simple test will help
avoid regression. It also highlights that the feature does not work on remote
iframe when Fission is enabled, which is why we tag it as `fail-if = fission`.

Differential Revision: https://phabricator.services.mozilla.com/D98047

UltraBlame original commit: 415c575e1ed0d02c1c9ff584ef137e1115af76b4