Bug 1668919 - try to make browser_tab_touch_events.js less racy, r=edgar
As far as I see the test is racy. It adds event listener to a window object, which reload() will replace.
So if the load event for non-transient about:blank gets chance to fire, then the test passes without the patch.
The main thing the patch does is to listen for the load event on the iframe element, since that object isn't
replaced.

The intermittent failure happens quite rarely and not locally, so the patch is based on code inspection.
Locally the test was run successfully 100 times both with opt and debug builds.

https://treeherder.mozilla.org/#/jobs?repo=try&revision=8b3c65c89c55cd16bf57cf0b78db9ee99ba23df2

Differential Revision: https://phabricator.services.mozilla.com/D98305

UltraBlame original commit: 4ab108ce38da756416b39c5c05f1a4cb1448f30f