Bug 1679986 - Add assertions to check that we don't leak arenas r=sfink
The main purpose of this is to assert that ArenaList objects are empty on
destruction, i.e. we're not accidentally leaking arenas.

This makes it so that we always move the contents of an ArenaList, which clears
the original list. I also removed a couple of unused methods.

Differential Revision: https://phabricator.services.mozilla.com/D98303

UltraBlame original commit: 507ea8573da02719ce2306b15da9b3928669c079