Bug 1677094 - Assert if we lose WebRender if forced via MOZ_WEBRENDER env var. r=jrmuizel
We use MOZ_WEBRENDER to force WebRender on in our testing
infrastructure. We may silently fallback to basic during our tests if we
encounter an error and the test may pass as a result. It would be best
if we asserted we don't lose WebRender while testing.

Differential Revision: https://phabricator.services.mozilla.com/D97004

UltraBlame original commit: b6eb6c19057d8ace54983a22b0daca46cf890564