Bug 1678015 - Expose assertive live region attributes for ARIA alerts on Mac, r=Jamie,eeejay
This is done via an ifdef in the ARIAMap. Provisions, however, are there to simply remove the ifdef and if clauses from here and the test file to enable this on other platforms once alert double-speaking and other unwanted side effects are fixed on said platforms.

Differential Revision: https://phabricator.services.mozilla.com/D97584

UltraBlame original commit: 5081ce938a4b0ebfa9d36d989b5555eebbef1acd