Bug 1657404 - Gamepad: Make a singleton to hold monitoring observers r=handyman
Currently, the list of monitoring observers is stored in the
GamepadPlatformService. But it's possible for testing to start before an
event channel has been created, or to exist longer. That could result in the
GamepadPlatformManager being created-destroyed multiple times along with
this list.

Probably the simplest thing to do here is just have the list be its own
indepedent entity

Differential Revision: https://phabricator.services.mozilla.com/D96663

UltraBlame original commit: 349854c5ad56f0885a5b84ba80f881b9ac9900ff