Bug 1679555 [wpt PR 26661] - Fix a few flaky test in css-paint-api/, a=testonly
Automatic update from web-platform-tests
Fix a few flaky test in css-paint-api/

The problem of these tests is that we do a takeScreenShot in the
rAF, and a rAF doesn't guarantee that the animation is ready, so
it can be easily flaky given that the animations are running on
the compositor thread.

This CL fixes it so that it calls takeScreenShot when the animation
ready promise is resolved, and the makes sure that the animation
already begun.

Bug: 1133821
Change-Id: I4fae6ccb9fe992316831d4d2013e1aa32eb52c82
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2562728
Reviewed-by: Kevin Ellis <keverschromium.org>
Commit-Queue: Xida Chen <xidachenchromium.org>
Cr-Commit-Position: refs/heads/master{#831650}

--

wpt-commits: 130f163afe563f8d858e960570bfdd9dc02cdbae
wpt-pr: 26661

UltraBlame original commit: a0b6c231b383d540f085a547059ac84f0d511dae