Bug 1678757 - [devtools] Add missing errors in Error previewer. r=bomsy.
We weren't handling the error exception object in the Error previewer for
some error classes, and as a result we'd show an undefined grip in the console,
instead of a proper error message.
A test is added to check that we do print wasm errors as expected.

Differential Revision: https://phabricator.services.mozilla.com/D98469

UltraBlame original commit: 85d1fafd696aadc3b5f53c79b918c2ebdf48dcb7