Bug 1677929 - Extract the 'root' parts of NotifyPipelineRendered into a new method called NotifyDidRender. r=mattwoodrow
This reduces some code duplication, and makes it clear which things we want to
do only once per window, and which things we want to do once per pipeline / CompositorBridge.

This change affects the ordering of messages, but hopefully not in a way that
anybody was relying on. Specifically, the image notifications are now sent before
the NotifyDidPaint for the root CompositorBridge is sent.

Differential Revision: https://phabricator.services.mozilla.com/D97533

UltraBlame original commit: 7eb45ae1b6c6f687ced992f233f2a909388b4d96