Bug 1673497: Part 15: Remove vestigial support for deferred call stubs r=jandem
This code was originally added to support ConstStringSplit. Later, we removed ConstStringSplit because it didn't do anything useful, so this code is dead.

(This patch originally simplified some of the bailout work. Now that we're intending to eagerly invalidate warp scripts at the beginning of the fallback, it's just a driveby fix.)

Differential Revision: https://phabricator.services.mozilla.com/D99164

UltraBlame original commit: e6182530443fc86c07a89b796c4fda8c872d092f