Bug 1678542 - Optimize branchTestPtr on ARM64.  r=nbp
branchTestPtr can be optimized in the manner of branchTest32, using CBZ/CBNZ
when the inputs are the same and we're testing for zero/nonzero.

Also tidy up the logic of branchTest32 to make it slightly easier to understand.

Differential Revision: https://phabricator.services.mozilla.com/D97770

UltraBlame original commit: d84189f9e7338ff7cfed22e02969d779e73cac23