Bug 1572654 - [devtools] Restore active flexbox highlighter on page refresh r=jdescottes
Depends on D96730

This patch refactors slightly the way flexbox highlighters are restored when the page is refreshed while a highlighter is active.
To account for Fission, we store the `browsingContextID` of the target in addition to the node's unique selector, in order to match again the node after refresh and show a flexbox highlighter for it.

The events emitted are renamed to more generic versions "highlighter-restored" and "highlighter-discarded". They contain the highlighter type in the event data payload. Tests are updated accordingly.

Also made slight changes to the grid highlighter restore logic and corresponding tests in order to use the new generic events and test helpers. The bulk of the refactoring for grid highlighters will happen in [Bug 1572652](https://bugzilla.mozilla.org/show_bug.cgi?id=1572652)

Differential Revision: https://phabricator.services.mozilla.com/D97660

UltraBlame original commit: f7509696a6d7084263b78339f4c835d5e3e75c20