Bug 1637658 - P1. Only use ImageRect if explicitly set. r=mattwoodrow,pehrsons
The concept of ImageRect only exists with the WebM container. It defines a rectangle that will be used to crop the decoded image.

Initially (and as still indicated in the comment for VideoInfo.mImageRect) that field was only ever initialised by the WebMDemuxer.
So we now make this an optional value that needs to be explicitly initialised to be active, and only do so in the WebMDemuxer.

Differential Revision: https://phabricator.services.mozilla.com/D99601

UltraBlame original commit: aa3a0821281ac95c324e712fd98ce346ed59e896