Bug 1680602 - Align stack pointer, and clear alignment area, before calling GeneratorResume r=iain
This doesn't matter at the moment, but once we can Warp compile generators we
will need the JIT stack to be JitStack aligned. As well, we will need
BaselineBailouts to understand that one of the possible stack frames could be
the BaselineFrame created by a generator body OSR'ing into Warp.

We need to zero out the area for alignment, as during GC it can be inspected
assuming it contains valid values.

Differential Revision: https://phabricator.services.mozilla.com/D98759

UltraBlame original commit: 038847f6366ca5424214300b2fff8a38e70fcfc7