Bug 1577530 - Cache theming values from nsTextControlFrame correctly. r=heycam
The issue is that nsTextControlFrame might have a different used padding
then the editing root block and thus the editing root's scrolled
content, etc, because its padding doesn't match the style.

This causes TextOverflow code to miscompute the content area size, which
is sadness and causes the overflow markers to go over the padding box.

The REFLOW_ROOT condition was introduced in bug 157846 and only
partially takes care of the issue. This fix supersedes it, so we can
remove it.

Differential Revision: https://phabricator.services.mozilla.com/D99838

UltraBlame original commit: fd07538e12316562b6479e77af02c7b2b0f94d43