Bug 1682606 - Ensure mach wpt-update has the correct sys.path, r=denschub
The code in the frontend was trying to do its own path manipulation, which then broke
with the pywebsocket changes. It's generally a better idea to just load localpaths.py
so that all the horribleness of the setup is contained in that one place.

Differential Revision: https://phabricator.services.mozilla.com/D99930

UltraBlame original commit: e4a8dbedad2a382951bd7acc0cc8a7e4f301130b