Bug 1683029 - Reintroduce f() and f_1(). r=aosmond
This fixes a subtle bug in the original c code where the results of f_1()
were not properly bracketed causing only the second part of the ternary
to be multiplied by the whitepoint.

I tried but was not able to create a test that showed the difference.

Differential Revision: https://phabricator.services.mozilla.com/D99970

UltraBlame original commit: 099856a46d6b2671ab23739c35b6772059afa10e