Bug 1683426. Assert in DManipEventHandler::OnContentUpdated that SendPinch returns true if mShouldSendPinchStart is true. r=botond
Otherwise we wouldn't send a pinch start. This should never be hit because if we get here the current scale should not be fuzzy equal to the previous scale so SendPinch should compute a non-zero deltaY and send an event and return true.

Differential Revision: https://phabricator.services.mozilla.com/D100170

UltraBlame original commit: 826fe8a15768483d77c81d2438c2d5d0a04c7992