Bug 1675615 - Move WR gpu cache logic into a renderer submodule r=gw
this is probably the last of the low-hanging fruits in renderer submodules.
I think it would be very useful to try isolating the scene building stuff in a similar way.
So in the end, the root `src` should only contain things that are used by multiple stages of WR.

Differential Revision: https://phabricator.services.mozilla.com/D100263

UltraBlame original commit: 3b76bd715b736f9946560bbe0f166b4303c50b4f