Bug 1678651 - Fixed sampler's detection of too-big stacks - r=canaltinova
The profiler sampler was using the "cleared block count" as indication that some data could not be stored.
But this was incorrect, because cleared blocks only account for blocks in chunks that have been destroyed or recycled, but this missed data that could not be stored due to lack of space.
In particular, `ProfileBufferChunkManagerSingle` only provides one chunk, so when it gets full, a new chunk cannot be allocated, and the data is lost; but the "cleared block count" is left unchanged (the one chunk is only released, but not destroyed nor recycled).

Now the sampler correctly looks at the "failed puts bytes", and can report precisely how many bytes could not be stored.

Differential Revision: https://phabricator.services.mozilla.com/D99978

UltraBlame original commit: ba92771a968a4d8226c3fb5ee53233085b263a19