Bug 1683980 - avoid producing more clipped vertexes than we can support. r=jrmuizel
This ensures that in the worst case in a release build we don't accidentally overflow
any vertex buffers as a result of floating point precision. Instead, we trade off for
rendering correctness by choosing just not to render the perspective quad since we can't
properly clip it. This still leaves in an assert for debugging so that we can go back
and figure out potentially why some case is triggering this.

Differential Revision: https://phabricator.services.mozilla.com/D100373

UltraBlame original commit: d7b25f7e2eefa4ea6f8209e7b2753701e81144c5