Bug 1684352 - Don't assume that there's a vsync dispatcher handy when observing. r=rmader
My guess is that if Observe() gets called too early, we can receive the
message before updating the vsync source the first time, which would
trigger this crash.

With this patch, we handle it nicely and UpdateVsyncSource will
re-observe as needed.

It's not 100% clear to me if this is actually the issue (and thus
whether this patch is the right solution), since I haven't been able to
repro locally, though. There might be a more subtle race.

Differential Revision: https://phabricator.services.mozilla.com/D100501

UltraBlame original commit: c39107f8d12aa8b2ef31ebf83e0e14839d618789