Bug 1680329 - Add AXTextStateChangeType to selection events. r=morgan
It seems that VoiceOver depends on that property to know if a selection is be extended or moved. This manifests in several places in Slack when the user combines VO navigation with Slack's keyboard navigation.

Also changed where we retrieve the editable ancestor to simplify the notification dispatching block.

Differential Revision: https://phabricator.services.mozilla.com/D100452

UltraBlame original commit: 5712fdcdf04f0573e30c5fcf14515fd27205114c