Bug 1655052 - ScopeContext now only checks this-env when it will be used. r=arai
Instead of generating bogus values inspecting the scope chain when the script
will provide a this-environment, we leave at default based on if the
InheritThis flag is passed. This is used for arrow-function and evals (as the
spec requires).

Differential Revision: https://phabricator.services.mozilla.com/D100848

UltraBlame original commit: 154dd55151c389145a4e211ebc625e84d2ad9b3e