Bug 1685078 - Fix test_bullet_animation.html. r=tnikkel
The current test relies on the explicit decoding that we start
triggering unconditionally right now, and which my previous patch
removes:

  https://searchfox.org/mozilla-central/rev/519f913527b0d9d5097d290d5731cff6b2991fe0/layout/generic/nsBulletFrame.cpp#1065

Remove the text so that the animation polling stuff detects when the
image loads (otherwise it sees a non-blank paint and starts thinking
that stuff is loaded, taking the reference snapshot too early).

Similarly, we need to set the color to transparent so that we don't get the
bullet that we draw as a fallback while the image is loading.

Differential Revision: https://phabricator.services.mozilla.com/D101034

UltraBlame original commit: 0da6fc5419b3c4f6dbb15cc349dc54a3153ff1e6