Bug 1682623 - Enable Warp compiling generators and async functions by default r=iain
We seem to have not had a further influx of fuzzing issues on this feature,
and my measurements of performance have shown no notable degradations (and a
few benchmark benefits), so I think it's time to let this ride the trains.

If we do this shortly, we'll have a few weeks of coverage on nightly before
soft freeze.

Differential Revision: https://phabricator.services.mozilla.com/D101083

UltraBlame original commit: 8924f5fc3c4b03019d010af3ac40aaa74606f3ec