Bug 1557247 - [devtools] Share jest config and fixtures across panels. r=ladybenko.
This creates a new devtools/client/shared/test-helpers folder in which we add
a share jest config, which define a `moduleNameMapper` option with all the
different fixtures that were used in the codebase.

The fixtures were moved from different panels to a jest-fixtures folder and
the one that were left unused are removed.
Some fixtures file needed to be modified to satisfy eslint rules.

Differential Revision: https://phabricator.services.mozilla.com/D99470

UltraBlame original commit: e901c527f766174bda61cdc289ad4484c671d03c