Bug 1357365 - part 4: Make `TypeInState` stop keeping link style when selection change command is performed, but it didn't change selection r=m_kato,edgar
When caret is not moved but selection change command is fired, it means that
the caret is already set to start or end of editing host.  In this case, we
should stop keeping link style for new inserting content because otherwise,
user cannot exit from the first or last link with arrow keys.

Differential Revision: https://phabricator.services.mozilla.com/D101003

UltraBlame original commit: 642a9425089b70d7b8992a43d77573ea1eb13d80