Bug 1686272 - jit_test: Clone tests when using --repeat option. r=tcampbell
The repeat option of `jit_tests.py` was interpreted as a plain copy of the
object pointer. Therefore the `xdr_annotate` function, first annotated the first
test with an `encode` mode for the self hosted code, but later overwrote it with
`decode` as the rests of the tests are annotated to make use of the serialized
self hosted stencil.

This change make use of the `copy` function of the test class to avoid sharing
the same instance between the first run and subsequent runs.

Differential Revision: https://phabricator.services.mozilla.com/D101755

UltraBlame original commit: 434b1d0b2971e281ba9306e66733c04b0a7078d5