Bug 1685985 - [devtools] Handle error count button like other toolbox buttons. r=jdescottes.
A button configuration is now created from _buildButtons so the error count button
is part of the `toolbarButtons` property on the toolbox.
An `updateErrorCountButton` updates the property of the button, and we call it
from the function that updates the error count, from which we also trigger a new
render for the toolbar.

This way updating the error count will go through the overflow logic we have in
the toolbar (to show the chevron or not).

A "side effect" of this is that an item is automatically created in the options
panel to disable the error count button. In order for it to work properly, we
need to add a preference to store its enabled state, as well as a specific
localized string for the input label.

browser_toolbox_error_count is updated to ensure that the error button isn't
displayed if it was disabled in the option panel, even if new errors are emitted.

browser_toolbox_options_disable_buttons was updated to reflect the changes made
to the error count button.

Differential Revision: https://phabricator.services.mozilla.com/D101569

UltraBlame original commit: 3ff621cea4dcd1b0c87391f689d272deea70c4b2