Bug 1686036 - [devtools] Avoid error count icon unnecessary render. r=jdescottes.
In `setErrorCount`, we used to call setToolboxButtons unconditionally, which means
that it was called each time a resource was added or updated.
This was very expensive and would cause some slowness, especially on pages with
a lot of network requests, as revealed by DAMP tests.

In this patch, we don't call setToolboxButtons if the error count did not change.
On top of that, we add a throttle mechanism so the function can't be called more
than twice per second.

Differential Revision: https://phabricator.services.mozilla.com/D102136

UltraBlame original commit: b300d2dfebc46734e7780a2a1a9063203804b34d