Bug 1686758 - Properly guard against redundant moves. r=jseward
The == and != operators on float registers take into account the type
of the register, so when we compare registers we must coerce to
a common type for the comparison to be meaningful.

I considered moving the guard into moveFloat32 and moveDouble, but
this problem seems to be limited to SIMD code, so I judged that
a local fix was best.

Differential Revision: https://phabricator.services.mozilla.com/D101786

UltraBlame original commit: 140c3473408fc1297f27381364c6e351a2962fed